// src/app/api-services/models/jr-user-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface JrUserRequest {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  role: 'ROLE_PLATFORM_ADMIN' | 'ROLE_COMPANY_ADMIN' | 'ROLE_ADMINISTRATOR' | 'ROLE_SALES_OPERATOR' | 'ROLE_USER';
  username: string;
}